Change what Case and Default accepts as a parent
It used to be like everywhere else, everything implementing Node
will work. But since these two code blocks can be only in a Switch,
it is correct to add a restriction, less space for a mistake.

Also, Switch has a Condition, which will be used later of for that
equality checks.

Updates #8