Move For construction into the package "lang"
In "p" it was repeated three times and it did not look well. I will
probably make construction even stronger, accepting conditions and
more.

Also "f" renamed to "lf" to make it consistent, it is a shortcut
for "lang.For" and it avoids name collision in the For part of
the code, where I use "f" for typecasted Node to For.