Add function which turns statement to node list
This is a common thing for me to do. Because conditions and cycles
work with blocks, single statements are forbidden, functions work
with arrays of nodes.

Check if node is a simple statement or a statement list is so essential,
that I am shocked I did not make this change commits before.