Turn PostInc and PostDec into statements
It caused some issues with expression creation, but now the code
generating is in the correct Go-way.

I will probably update this again, to allow {Post, Dec} as expressions,
but that will be done by cheating later on no easy way to make it an
expression, especially with the incoming introduction of the interfaces.