Add some error handling to the code
Until now, no function returned error. It is about time I tried to do something
with it. I will be able to debug code and test if it fails how it should fail.
Panicing about everything is not the best way to do things.

Also updates the quote when something is not caught by expression solver,
prefix where the issue happened is useful.