Remove function HasVariable from the Node interface
HasVariable turned out to be something very specific, useful only
for Blocks, these define variable with method DefineVariable.

Parent and SetParent is now ugly, I might do something about it later on.
I want to keep it, to make sure everything is linked properly and I can
walk to root of the function from every node. It can turn out to be useful
when I will start with "continue NUMBER" and other not so nice constructs.

I also finally decided what I think about Parent in Node. It just has to
be here, to be paired up with the setter.