Resolve issues with keywords
I have to avoid 25 of the reserved words in the Go. PHP does not
have this issue, they hide every conflict behind a dollar.

It looks like functions does not have to be solved, PHP has the
same issue as Go, function named "for" cannot exist and swapping
func to function is already implemented.

Closes #3