Draft associative array to the std library
With this, I will be able to work with the arrays almost the same
way as the PHP does, that means indexing using string/bool/float etc.,
keeping order in which the items were inserted.

Note that this is not yet used in the parser.

Tests should be added, this is the third or fourth time I tried to do
something like this.

Updates #15