Add result comparision
This will be another way how to test that the code works, finding
out if the output is the same as the PHP one.

Most of the stuff can probably be caught by comparing expected
source code and generated one, but this can get complicated when
trying to resolve bigger codes, behaviour with dynamic content
like database connection etc.

This also has one advantage of doing benchmarks later on; but this
is not coming any time soon.

GlobalContext had to be changed, not every example uses my small
standard library.