Work with variable visibility
In Go everything starting with the uppercased letter is a public
function/variable. This change makes sure that every variable which
is not in the function (= is global), is uppercased and thanks to
that is global, behaves the same way in Go as in PHP.

With this change it will be possible to translate multiple files
without any issues (note that issues cannot appear right now,
because I do not know what a namespace is), everything important
will be stored in the globalContext, which will be able to answer
inform everybody that HasVariable(name) is here.

The solution is exceptionaly ugly and some things are being duplicated,
stored in global variables and more. This makes testing and adding
new functionality very hard. After finishing the array and server part
of the compiler, big refactoring will appear.

Updates #2