Add generator for arrays
Arrays are strongly typed and because this language does not have
generics, I have to do something extra to solve this issue.

One of the options is generating; every type will have its own array.
Doing this manually would be really annoying after some time, so
the generator was added - it is not doing anything complicated, it just
puts type name to the correct places.

IntArray was re-generated and strings added.

Update #15