Change how type switches are used
I was using them like this before because I was not so familiar
with them, did not know I can assign in the argument part of the
switch and know type later.

See https://tour.golang.org/methods/16

Some parts of the code can be more magical because I do not
use cool names like "issets" and others. Instead of that I have
generic "e" for "expression" and so on. I might change it later
on, but if the code is not understandable it just means the
case is too big.

Also one switch is removed and replaced by simple cast with the
second return value signign how successful it was. With that
I got rid of one extra indentation and possibly a cleaner code,
now it reallt looks like there is only one possible option, not
N cases.