Fix issue with empty return statement
I wanted to fix an issue with undefined function, how they require given order
to be inserted and in some cases this issue cannot be resolved, like the one
in examples/32.php.

But because parser ignores it, there is no issue of this type. But later on I
will have to add this knowledge to the parser, otherwise I cannot resolve
varid argument count functions.