Implement if conditions with statement as a condition
Title is misleading, this is just resolving a gap between PHP
and Go, in Go assignments, {inc,dec}rements are statements and
because of that they cannot be used as a condition.

Now it works only with an assignment, methods around expressions
need some refactoring before I can implement more of these things.

Updates #7