Implement out of scope
This is the first working implementation of the out of scope issue
common in interpreted languages like PHP, Python, JS (var).

When it is necessary, variable is pulled out one of the previous
blocks. Finding assignment before block with the variable stops
this behaviour and everyhing works as it should. Let's call it
"gift for a high quality code, less interfaces".

Complex data types like std/array were not checked, probably does
not work. This issue will be solved later on when I will have
struct methods under control.

Interface might be changed later on, that out of scope variable is
used most of the time as 'true', only when creating assignment the
value is 'false'.

Note that 'case' and 'default' are not updated. There are two reasons
for that: I was lazy and I want to change them to use *Code instead.
Code associated with this variable definition is very ugly.

Closes #5