Implement default parameters
This is a very big change. I had to change how function.go works
with the functions and change its behaviour: instead of one function
for given keyword I have an array. Array contains every definition
of the tweaked function, they differ by their argument count. Or more
precisely, how many of them are omitted.

It discovered how badly I work with parser.funcs.Add, GlobalContext.Add
has to be called twice, because filling the function body is done with
a different reference.

PHP can do more with default parameters than I do. They are OK with gaps
in the definitions. This is something am not willing to transpile, it
has to be without gaps, without variadic count and defaults start from
the end of the argument list.

Closes #4