Change how {inc,dec}rements are handled
Previous solution had some flaws, for instance string could not be incremented.
This is the first solution for this issue, more of a draft. When type is a string,
it will use an extra function passed as an argument.

Next goal is to implement string increments how PHP knows them and turn these statements
to expressions if necessary.

Updates #13